Crate relm4_css

Static definitions for Adwaita’s style classes and colors.

Most of the functionality requires Adwaita being active in your application, though you may have luck using them with just GTK.

Available classes: https://gnome.pages.gitlab.gnome.org/libadwaita/doc/1.2/style-classes.html.

Available colors: https://gnome.pages.gitlab.gnome.org/libadwaita/doc/1.2/named-colors.html#palette-colors.

Docs of related crates: relm4 | relm4-macros | relm4-components | relm4-css | gtk4-rs | gtk-rs-core | libadwaita-rs | libpanel-rs

GitHub | Website | Book | Blog

Re-exports

pub use classes::*;

pub use colors::*;

Modules

classes

Adwaita’s CSS classes.

colors

Adwaita’s CSS colors.